+++ app/scripts/Log_clear.py
