+++ app/scripts/Stall_handler.py
